+++ CHANGELOG.md
@@ -24,16 +24,21 @@
 - Feedback: retrieving loved/hated tracks from library will also output library tagged tracks (if available), without duplicates by MBID.
 - Feedback: menu entries will be disabled when selection count is higher than the API recommendations.
 - Feedback: in case sending feedback to server fails (usually due to rate limits), connection will be retried once a few ms later. Any error will be reported via console/popups.
+- Feedback: only updates tracks on server if the current feedback differs, this way the timestamp does not get changed.
+- Feedback: improvements when handling tracks without MBIDs.
 - Playlists: matches on library -for playlist creation- are now preferred by higher rating and not live tracks (if possible).
 - Playlists: optimizations finding tracks on library.
+- Playlists: standardized playlists names, similar to playlist created by [Last.fm-SMP](https://regorxxx.github.io/foobar2000-SMP.github.io/scripts/lastfm-smp/).
 - ListenBrainz: key for the token is cached during the same session.
 - ListenBrainz: user name is cached during same session.
+- ListenBrainz: updated getFeedback method (with POST), without limits on data size.
 - UI: all reports are now formatted as tabulated tables, making them easier to read.
 - UI: tooltip now reports if user playlists recommendations are available and feedback for the selected track.
 - UI: button is now animated during all asynchronous tasks (including daily jams retrieval).
 ### Removed
 ### Fixed
 - Feedback: retrieval of tracks with feedback on server was incomplete due to API limits of number of items get per request. Now requests are paginated, untill all results are get.
+- Lookup for missing MBIDs setting was not honored (always active).